src/producer/requirements.txt
azure-eventhub
pandas